requirements.txt
nose
python-dateutil
isoweek